+++ src/main/java/com/buuz135/functionalstorage/block/tile/DrawerTile.java
@@ -70,7 +70,8 @@ public class DrawerTile extends ItemControllableDrawerTile<DrawerTile> {
                 type.getSlots(),
                 type.getSlotPosition(),
                 integer -> getHandler().getStackInSlot(integer),
-                integer -> getHandler().getSlotLimit(integer)
+                integer -> getHandler().getSlotLimit(integer),
+                integer -> getHandler().getStoredStacks().get(integer).getStack()
         ));
     }